+++ spec/spec_helper.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-ENV["DISABLE_TYPELIZER"] = "false"
+ENV["TYPELIZER"] = "true"
 ENV["RAILS_ENV"] = "test"
 require File.expand_path("app/config/environment", __dir__)